+++ src/main/java/org/jabref/logic/importer/fileformat/BibTeXMLImporter.java
@@ -25,7 +25,7 @@ import org.jabref.logic.importer.fileformat.bibtexml.Entry;
 import org.jabref.logic.importer.fileformat.bibtexml.File;
 import org.jabref.logic.importer.fileformat.bibtexml.Inbook;
 import org.jabref.logic.importer.fileformat.bibtexml.Incollection;
-import org.jabref.logic.util.FileType;
+import org.jabref.logic.util.StandardFileType;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibtexEntryTypes;
 import org.jabref.model.entry.FieldName;
@@ -54,8 +54,8 @@ public class BibTeXMLImporter extends Importer {
     }
 
     @Override
-    public FileType getFileType() {
-        return FileType.BIBTEXML;
+    public StandardFileType getFileType() {
+        return StandardFileType.BIBTEXML;
     }
 
     @Override