+++ src/main/java/org/jabref/preferences/CustomExportList.java
@@ -13,6 +13,7 @@ import org.jabref.logic.exporter.TemplateExporter;
 import org.jabref.logic.journals.JournalAbbreviationLoader;
 import org.jabref.logic.layout.LayoutFormatterPreferences;
 import org.jabref.logic.util.FileType;
+import org.jabref.logic.util.StandardFileType;
 
 import ca.odell.glazedlists.BasicEventList;
 import ca.odell.glazedlists.EventList;
@@ -88,12 +89,16 @@ public class CustomExportList {
             SavePreferences savePreferences) {
 
         String lfFileName;
-        if (extension.endsWith(".layout")) {
+        if (filename.endsWith(".layout")) {
             lfFileName = filename.substring(0, filename.length() - ".layout".length());
         } else {
             lfFileName = filename;
         }
-        TemplateExporter format = new TemplateExporter(exporterName, filename, lfFileName, null, FileType.parse(extension), layoutPreferences,
+        if (extension.contains(".")) {
+            extension = extension.substring(extension.indexOf('.') + 1, extension.length());
+        }
+        FileType fileType = StandardFileType.newFileType(extension);
+        TemplateExporter format = new TemplateExporter(exporterName, exporterName, lfFileName, null, fileType, layoutPreferences,
                 savePreferences);
         format.setCustomExport(true);
         return Optional.of(format);