+++ src/test/java/org/jabref/logic/importer/fileformat/OvidImporterTest.java
@@ -13,7 +13,7 @@ import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 import org.jabref.logic.bibtex.BibEntryAssert;
-import org.jabref.logic.util.FileType;
+import org.jabref.logic.util.StandardFileType;
 import org.jabref.model.entry.BibEntry;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -59,7 +59,7 @@ public class OvidImporterTest {
 
     @Test
     public void testsGetExtensions() {
-        assertEquals(FileType.OVID, importer.getFileType());
+        assertEquals(StandardFileType.TXT, importer.getFileType());
     }
 
     @Test